Moved DebugValueList to it's own file
